Function to calculate first coalescence copying vector for each population specified in poplabels file.

This function calculates the proportion of the genome where the first coalescence is with group x. It returns a version where coalescence with own group is counted, and one where the own group is excluded.

ExpPaintingProfile(file_anc, file_mut, poplabels, pops = NULL, chrs = NULL)

Arguments

file_anc

Filename of anc file. If chrs is specified, this should only be the prefix, resulting in filenames of ${file_anc}_chr${chr}.anc(.gz).

file_mut

Filename of mut file. If chrs is specified, this should only be the prefix, resulting in filenames of ${file_anc}_chr${chr}.anc(.gz).

poplabels

Filename of poplabels file

pops

(Optional) Populations for which data should be extracted. Names need to match the second column of the poplabels file

chrs

(Optional) Vector of chromosome IDs

Value

2d array of dimension #groups x #groups.

Examples

file_anc  <- system.file("sim/msprime_ad0.8_split250_1_chr1.anc.gz", package = "twigstats")
file_mut  <- system.file("sim/msprime_ad0.8_split250_1_chr1.mut.gz", package = "twigstats")
poplabels <- system.file("sim/msprime_ad0.8_split250_1.poplabels", package = "twigstats")

#Calculate f2s between all pairs of populations
nn_res <- ExpPaintingProfile(file_anc, file_mut, poplabels)
if(require(lsei)){
  pnnls(t(nn_res[c("P2","P3"),c("P1","P2","P3","P4"),1]),(nn_res["PX",c("P1","P2","P3","P4"),1]), sum = 1)$x
}
#> Loading required package: lsei
#> [1] 0.2397118 0.7602882